requirements.txt
git+https://github.com/mccoderpy/discord.py-message-components.git@developer
python-dotenv
aiomysql
aiofiles
pillow

